Refactoring package hungryman to wherefordinner in availability app.
Adding profiles for cloud native events only.

Adding support for postgres database.
