+++ where-for-dinner/where-for-dinner-availability/pom.xml
@@ -5,17 +5,17 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.7.4</version>
+		<version>2.7.5</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
 	<groupId>com.example</groupId>
 	<artifactId>where-for-dinner-availability</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>where-for-dinner-availability</name>
-	<description>Hugryman service for storing and retriving establishment reservation openings.</description>
+	<description>Where For Dinner service for storing and retriving establishment reservation openings.</description>
 	<properties>
 		<java.version>11</java.version>
-		<spring-cloud.version>2021.0.3</spring-cloud.version>
+		<spring-cloud.version>2021.0.4</spring-cloud.version>
 		<spring-cloud-bindings.version>1.10.0</spring-cloud-bindings.version>
 		<snakeyaml.version>1.32</snakeyaml.version>
 	</properties>
@@ -32,14 +32,6 @@
 		    <groupId>org.springframework.boot</groupId>
 		    <artifactId>spring-boot-starter-oauth2-resource-server</artifactId>
 		</dependency>			
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-stream</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-stream-binder-rabbit</artifactId>
-		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-function-webflux</artifactId>
@@ -67,7 +59,11 @@
         <dependency>
             <groupId>org.mariadb</groupId>
             <artifactId>r2dbc-mariadb</artifactId>
-        </dependency>				
+        </dependency>	
+        <dependency>
+            <groupId>org.postgresql</groupId>
+            <artifactId>r2dbc-postgresql</artifactId>
+        </dependency>			
 	    <dependency>
 	       <groupId>io.r2dbc</groupId>
 	       <artifactId>r2dbc-h2</artifactId>
@@ -86,13 +82,6 @@
 			<groupId>io.projectreactor</groupId>
 			<artifactId>reactor-test</artifactId>
 			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-stream</artifactId>
-			<scope>test</scope>
-			<classifier>test-binder</classifier>
-			<type>test-jar</type>
 		</dependency>
 	</dependencies>
 	<dependencyManagement>
@@ -138,5 +127,38 @@
 			</plugin>
 		</plugins>
 	</build>
+    <profiles>
+       <profile>
+           <id>streams</id>
+           <activation>
+              <activeByDefault>true</activeByDefault>
+           </activation> 
+           <dependencies>
+			   <dependency>
+		            <groupId>org.springframework.cloud</groupId>
+		            <artifactId>spring-cloud-stream</artifactId>
+		       </dependency>	
+			   <dependency>
+					<groupId>org.springframework.cloud</groupId>
+					<artifactId>spring-cloud-stream-binder-rabbit</artifactId>
+			   </dependency>
+			   <dependency>
+					<groupId>org.springframework.cloud</groupId>
+					<artifactId>spring-cloud-stream</artifactId>
+					<scope>test</scope>
+					<classifier>test-binder</classifier>
+					<type>test-jar</type>
+			   </dependency>	   		   		
+		   </dependencies>
+       </profile>
+       <profile>
+           <!-- Stubbed for native image build -->
+           <id>native</id>
+           <dependencies>
 
+		   </dependencies>
+           <build>
+           </build>
+       </profile>
+    </profiles>
 </project>
