+++ where-for-dinner/where-for-dinner-availability/src/main/java/com/java/example/tanzu/wherefordinner/WhereForDinnerAvailabilityApplication.java
@@ -1,13 +1,13 @@
-package com.java.example.tanzu.hungryman;
+package com.java.example.tanzu.wherefordinner;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class HungrymanAvailabilityApplication {
+public class WhereForDinnerAvailabilityApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(HungrymanAvailabilityApplication.class, args);
+		SpringApplication.run(WhereForDinnerAvailabilityApplication.class, args);
 	}
 
 }
