+++ where-for-dinner/where-for-dinner-availability/src/main/java/com/java/example/tanzu/wherefordinner/repository/AvailabilityWindowRepository.java
@@ -1,7 +1,8 @@
-package com.java.example.tanzu.hungryman.repository;
+package com.java.example.tanzu.wherefordinner.repository;
 
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-import com.java.example.tanzu.hungryman.entity.AvailabilityWindow;
+
+import com.java.example.tanzu.wherefordinner.entity.AvailabilityWindow;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
