+++ where-for-dinner/where-for-dinner-availability/src/main/resources/application.yaml
@@ -7,6 +7,7 @@ server:
   
 spring: 
   sql.init.mode: ALWAYS
+  sql.init.platform: h2
   
   cloud: 
     stream:
@@ -22,14 +23,23 @@ spring:
             backOffMaxInterval: 60000 
             
 springdoc:
-  packages-to-scan: com.java.example.tanzu.hungryman.resources
+  packages-to-scan: com.java.example.tanzu.wherefordinner.resources
   show-spring-cloud-functions: false
   
 ---
 spring:
   config.activate.on-profile: mysql
 
+  sql.init.platform: mysql
+
   r2dbc: 
     #Default properties to work with MariaDB driver
     properties.sslMode: TRUST  
     properties.tlsVersion: TLSv1.2  
+    
+---
+
+spring:
+  config.activate.on-profile: postgres
+  
+  sql.init.platform: postgresql
