+++ where-for-dinner/where-for-dinner-availability/src/main/resources/schema-h2.sql
